feat: add tabs and workspace setup on project creation
