+++ src/renderer/components/_templates/[workspace]/main-content.tsx
@@ -1,15 +1,39 @@
 import { useOpenPLCStore } from '@root/renderer/store'
+import { TabsProps } from '@root/renderer/store/slices'
+import { CreateEditorObjectFromTab } from '@root/renderer/store/slices/tabs/utils'
 import { cn } from '@root/utils'
-import { ComponentPropsWithoutRef } from 'react'
+import { ComponentPropsWithoutRef, useEffect } from 'react'
 
 type IWorkspaceMainContentProps = ComponentPropsWithoutRef<'div'>
 const WorkspaceMainContent = (props: IWorkspaceMainContentProps) => {
   const {
+    project: {
+      data: { pous },
+    },
     workspace: {
       systemConfigs: { OS },
     },
+    tabsActions: { updateTabs },
+    editorActions: { addModel, setEditor },
   } = useOpenPLCStore()
   const { children, ...res } = props
+
+  useEffect(() => {
+    const handleCreateTab = () => {
+      const tabToBeCreated: TabsProps = {
+        name: pous[0].data.name,
+        path: '/data/pous/program/main',
+        elementType: { type: 'program', language: pous[0].data.language },
+      }
+      updateTabs(tabToBeCreated)
+      const model = CreateEditorObjectFromTab(tabToBeCreated)
+      addModel(model)
+      setEditor(model)
+      return
+    }
+    handleCreateTab()
+  }, [])
+
   return (
     <div
       className={cn(
